Add the library screen to TV app
