+++ Jetcaster/tv-app/src/main/java/com/example/jetcaster/tv/ui/library/LibraryScreenViewModel.kt
@@ -1,0 +1,99 @@
+/*
+ * Copyright 2024 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.jetcaster.tv.ui.library
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import com.example.jetcaster.core.data.database.model.PodcastWithExtraInfo
+import com.example.jetcaster.core.data.di.Graph
+import com.example.jetcaster.core.data.repository.EpisodeStore
+import com.example.jetcaster.core.data.repository.PodcastStore
+import com.example.jetcaster.core.data.repository.PodcastsRepository
+import com.example.jetcaster.tv.model.EpisodeList
+import com.example.jetcaster.tv.model.PodcastList
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.launch
+
+class LibraryScreenViewModel(
+    private val podcastsRepository: PodcastsRepository = Graph.podcastRepository,
+    private val episodeStore: EpisodeStore = Graph.episodeStore,
+    private val podcastStore: PodcastStore = Graph.podcastStore,
+) : ViewModel() {
+
+    private val followingPodcastListFlow = podcastStore.followedPodcastsSortedByLastEpisode().map {
+        PodcastList(it)
+    }
+
+    @OptIn(ExperimentalCoroutinesApi::class)
+    private val latestEpisodeListFlow = podcastStore
+        .followedPodcastsSortedByLastEpisode()
+        .flatMapLatest { podcastList ->
+            if (podcastList.isNotEmpty()) {
+                combine(podcastList.map { episodeStore.episodesInPodcast(it.podcast.uri, 1) }) {
+                    it.map { episodes ->
+                        episodes.first()
+                    }
+                }
+            } else {
+                flowOf(emptyList())
+            }
+        }.map {
+            EpisodeList(it)
+        }
+
+    val uiState =
+        combine(followingPodcastListFlow, latestEpisodeListFlow) { podcastList, episodeList ->
+            if (podcastList.isEmpty()) {
+                LibraryScreenUiState.NoSubscribedPodcast
+            } else {
+                LibraryScreenUiState.Ready(podcastList, episodeList)
+            }
+        }.stateIn(
+            viewModelScope,
+            SharingStarted.WhileSubscribed(5_000),
+            LibraryScreenUiState.Loading
+        )
+
+    fun unsubscribe(podcast: PodcastWithExtraInfo) {
+        if (podcast.isFollowed) {
+            viewModelScope.launch {
+                podcastStore.togglePodcastFollowed(podcast.podcast.uri)
+            }
+        }
+    }
+
+    init {
+        viewModelScope.launch {
+            podcastsRepository.updatePodcasts(false)
+        }
+    }
+}
+
+sealed interface LibraryScreenUiState {
+    data object Loading : LibraryScreenUiState
+    data object NoSubscribedPodcast : LibraryScreenUiState
+    data class Ready(
+        val subscribedPodcastList: PodcastList,
+        val latestEpisodeList: EpisodeList,
+    ) : LibraryScreenUiState
+}
